# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student Career Development and Placement Readiness System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Auth tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Jobs ─────────────────────────────────────────────────────────────────────
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interviews
      responses:
        "200":
          description: Interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: scheduleInterview
      tags: [interviews]
      summary: Schedule an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    patch:
      operationId: updateInterview
      tags: [interviews]
      summary: Update interview status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"

  # ─── Notifications ───────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: adminCreateUser
      tags: [admin]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user role / status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/departments:
    get:
      operationId: listDepartments
      tags: [admin]
      summary: List departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [admin]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /admin/departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [admin]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: adminDeleteDepartment
      tags: [admin]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, officer, admin]
        departmentId:
          type: ["integer", "null"]
        rollNo:
          type: ["string", "null"]
        batch:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, officer, admin]
        isActive:
          type: boolean
        departmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [student, officer, admin]
        isActive:
          type: boolean

    # ─── Student ──────────────────────────────────────────────────────────────
...
    Company:
      type: object
      required: [id, name, industry]
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
          type: string
        description:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        minCGPA:
          type: ["number", "null"]
        minCGPARequired:
          type: number
        maxBacklogs:
          type: integer
        requiredSkills:
          type: array
          items:
            type: string
        packageMin:
          type: ["number", "null"]
        packageMax:
          type: ["number", "null"]
        eligibleDepartments:
          type: array
          items:
            type: string
    CompanyInput:
      type: object
      required: [name, industry]
      properties:
        name:
          type: string
        industry:
          type: string
        description:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        minCGPARequired:
          type: number
        maxBacklogs:
          type: integer
        requiredSkills:
          type: array
          items:
            type: string
        packageMin:
          type: number
        packageMax:
          type: number
        eligibleDepartments:
          type: array
          items:
            type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        description:
          type: string
        website:
          type: string
        minCGPARequired:
          type: number
        maxBacklogs:
          type: integer
        requiredSkills:
          type: array
          items:
            type: string
        packageMin:
          type: number
        packageMax:
          type: number
        eligibleDepartments:
          type: array
          items:
            type: string

    # ─── Job Posting ──────────────────────────────────────────────────────────
...
    Interview:
      type: object
      required: [id, studentId, studentName, companyId, companyName, scheduledAt, type, status]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        jobId:
          type: ["integer", "null"]
        scheduledAt:
          type: string
        type:
          type: string
          enum: [technical, hr, group_discussion, aptitude, final]
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        venue:
          type: ["string", "null"]
        feedback:
          type: ["string", "null"]
        result:
          type: ["string", "null"]
    InterviewInput:
      type: object
      required: [studentId, companyId, scheduledAt, type]
      properties:
        studentId:
          type: integer
        companyId:
          type: integer
        jobId:
          type: integer
        scheduledAt:
          type: string
        type:
          type: string
          enum: [technical, hr, group_discussion, aptitude, final]
        venue:
          type: string
    InterviewUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, rescheduled]
        venue:
          type: string
        feedback:
          type: string
        result:
          type: string

    # ─── Notification ─────────────────────────────────────────────────────────
...
    Department:
      type: object
      required: [id, name, code]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        hodName:
          type: ["string", "null"]
        studentCount:
          type: integer
    DepartmentInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        hodName:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        hodName:
          type: string

    # ─── Admin Dashboard ──────────────────────────────────────────────────────
...