# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Student Career Development and Placement Readiness System API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  industry: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  minCGPA?: number | null;
  minCGPARequired?: number;
  maxBacklogs?: number;
  requiredSkills?: string[];
  /** @nullable */
  packageMin?: number | null;
  /** @nullable */
  packageMax?: number | null;
  eligibleDepartments?: string[];
}