# lib/api-zod/src/generated/types/departmentUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Student Career Development and Placement Readiness System API
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentUpdate {
  name?: string;
  code?: string;
  hodName?: string;
}