# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Student Career Development and Placement Readiness System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  isActive?: boolean;
  /** @nullable */
  departmentId?: number | null;
  createdAt: string;
}